move search_message to EventCommon
