+++ src/sentry/eventstore/models.py
@@ -1,19 +1,13 @@
 from __future__ import absolute_import
 
 import pytz
-import six
 
 from datetime import datetime
 from dateutil.parser import parse as parse_date
 
-from django.conf import settings
-from django.utils.encoding import force_text
-
-from sentry import eventtypes
 from sentry.models import EventCommon, EventDict
 from sentry.db.models import NodeData
 from sentry.snuba.events import Columns
-from sentry.utils.safe import trim
 
 
 def ref_func(x):
@@ -94,37 +88,6 @@ class Event(EventCommon):
         update this field in event_manager.save().
         """
         self._message = value
-
-    @property
-    def search_message(self):
-        """
-        The internal search_message attribute is only used for search purposes.
-        It adds a bunch of data from the metadata and the culprit.
-        """
-        data = self.data
-        culprit = self.culprit
-
-        event_metadata = self.get_event_metadata()
-
-        if event_metadata is None:
-            event_metadata = eventtypes.get(self.get_event_type())().get_metadata(self.data)
-
-        message = ""
-
-        if data.get("logentry"):
-            message += data["logentry"].get("formatted") or data["logentry"].get("message") or ""
-
-        if event_metadata:
-            for value in six.itervalues(event_metadata):
-                value_u = force_text(value, errors="replace")
-                if value_u not in message:
-                    message = u"{} {}".format(message, value_u)
-
-        if culprit and culprit not in message:
-            culprit_u = force_text(culprit, errors="replace")
-            message = u"{} {}".format(message, culprit_u)
-
-        return trim(message.strip(), settings.SENTRY_MAX_MESSAGE_LENGTH)
 
     @property
     def datetime(self):
